Extract PanProteome into own file.
