+++ src/proteomes/components/entry/Overview.tsx
@@ -1,12 +1,5 @@
 import { useMemo } from 'react';
-import {
-  Card,
-  InfoList,
-  ExternalLink,
-  LongNumber,
-  Loader,
-} from 'franklin-sites';
-import { SetRequired } from 'type-fest';
+import { Card, InfoList, ExternalLink, LongNumber } from 'franklin-sites';
 
 import HTMLHead from '../../../shared/components/HTMLHead';
 import TaxonomyView from '../../../shared/components/entry/TaxonomyView';
@@ -14,58 +7,21 @@ import { EntryTypeIcon } from '../../../shared/components/entry/EntryTypeIcon';
 import BuscoView from '../BuscoView';
 import BuscoLegend from '../BuscoLegend';
 import BuscoAbbr from '../BuscoAbbr';
-
-import useDataApi from '../../../shared/hooks/useDataApi';
+import { PanProteome } from './PanProteome';
 
 import parseDate from '../../../shared/utils/parseDate';
 import ftpUrls from '../../../shared/config/ftpUrls';
 import ProteomesColumnConfiguration, {
   ProteomesColumn,
 } from '../../config/ProteomesColumnConfiguration';
-import apiUrls from '../../../shared/config/apiUrls';
 
-import {
-  ProteomesAPIModel,
-  ProteomesUIModel,
-} from '../../adapters/proteomesConverter';
+import { ProteomesUIModel } from '../../adapters/proteomesConverter';
 
 import '../styles/overview.scss';
-import { Namespace } from '../../../shared/types/namespaces';
 
 type Item = {
   title: JSX.Element;
   content: JSX.Element;
-};
-
-type PanProteomeProps = SetRequired<
-  Pick<ProteomesUIModel, 'panproteome' | 'id' | 'taxonomy'>,
-  'panproteome'
->;
-
-export const PanProteome = ({
-  panproteome,
-  id,
-  taxonomy,
-}: PanProteomeProps) => {
-  const entryIsPanProteome = id === panproteome;
-  const { data: panProteomeData, loading } = useDataApi<ProteomesAPIModel>(
-    entryIsPanProteome ? null : apiUrls.entry(id, Namespace.proteomes)
-  );
-
-  if (loading) {
-    return <Loader />;
-  }
-
-  const name =
-    (entryIsPanProteome && taxonomy?.scientificName) ||
-    panProteomeData?.taxonomy.scientificName ||
-    panproteome;
-  return (
-    <>
-      {`This proteome is part of the ${name} pan proteome (`}
-      <a href={ftpUrls.panProteomes(panproteome)}>FASTA</a>)
-    </>
-  );
 };
 
 export const Overview = ({ data }: { data: ProteomesUIModel }) => {
